added interface for function hooking
